# openapi.yaml
...
openapi: 3.1.0
info:
  title: 大模型安全评测系统 API
  description: 大模型安全评测系统后端接口骨架。本服务默认监听本机，不包含登录、用户、角色权限和多租户。
  version: 1.0.0
paths:
...
  /api/v1/system-configs:
    get:
      tags:
      - 系统
      summary: List System Configs
      operationId: listSystemConfigs
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_SystemConfig__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/system-configs/{config_key}:
    patch:
      tags:
      - 系统
      summary: Update System Config
      operationId: updateSystemConfig
      security:
      - LocalToken: []
      parameters:
      - name: config_key
        in: path
        required: true
        schema:
          type: string
          title: Config Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfigUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SystemConfig_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/audit-logs:
    get:
      tags:
      - 系统
      summary: List Audit Logs
      operationId: listAuditLogs
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_AuditLog__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/evaluation-tasks:
    post:
      tags:
      - 评测任务
      summary: Create Evaluation Task
      operationId: createEvaluationTask
      security:
      - LocalToken: []
      parameters:
      - name: X-Operator
        in: header
        required: false
        schema:
          type: string
          default: local
          title: X-Operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationTaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EvaluationTask_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - 评测任务
      summary: List Evaluation Tasks
      operationId: listEvaluationTasks
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: benchmark_version_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Benchmark Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_EvaluationTask__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/status:
    get:
      tags:
      - 评测任务
      summary: Get Evaluation Task Status
      operationId: getEvaluationTaskStatus
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskStatusData_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/results:
    get:
      tags:
      - 评测任务
      summary: List Evaluation Results
      operationId: listEvaluationResults
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      - name: risk_level
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Level
      - name: needs_review
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Needs Review
      - name: model_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Model Id
      - name: attack_template_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Attack Template Id
      - name: risk_category_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Category Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_EvaluationResult__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/actions/start:
    post:
      tags:
      - 评测任务
      summary: Start Evaluation Task
      operationId: startEvaluationTask
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskActionData_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/actions/cancel:
    post:
      tags:
      - 评测任务
      summary: Cancel Evaluation Task
      operationId: cancelEvaluationTask
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskActionData_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/actions/retry-failed:
    post:
      tags:
      - 评测任务
      summary: Retry Failed Evaluation Cases
      operationId: retryFailedEvaluationCases
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RetryFailedData_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/statistics:
    get:
      tags:
      - 统计分析
      summary: Get Evaluation Task Statistics
      operationId: getEvaluationTaskStatistics
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Statistics_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/events:
    get:
      tags:
      - 评测任务
      summary: Stream Evaluation Task Events
      operationId: streamEvaluationTaskEvents
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_EvaluationTask_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/EvaluationTask'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[EvaluationTask]
...
    ApiResponse_PaginatedData_AuditLog__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_AuditLog_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[AuditLog]]
...
    ApiResponse_PaginatedData_EvaluationResult__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_EvaluationResult_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[EvaluationResult]]
    ApiResponse_PaginatedData_EvaluationTask__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_EvaluationTask_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[EvaluationTask]]
...
    ApiResponse_PaginatedData_SystemConfig__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_SystemConfig_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[SystemConfig]]
...
    ApiResponse_RetryFailedData_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/RetryFailedData'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[RetryFailedData]
...
    ApiResponse_Statistics_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/Statistics'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[Statistics]
...
    ApiResponse_SystemConfig_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/SystemConfig'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[SystemConfig]
    ApiResponse_TaskActionData_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/TaskActionData'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[TaskActionData]
    ApiResponse_TaskStatusData_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/TaskStatusData'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[TaskStatusData]
...
    EvaluationTask:
      properties:
        task_id:
          type: integer
          title: Task Id
        name:
          type: string
          title: Name
        benchmark_version_id:
          type: integer
          title: Benchmark Version Id
        risk_taxonomy_id:
          type: integer
          title: Risk Taxonomy Id
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Progress
          default: 0
        total_cases:
          type: integer
          title: Total Cases
          default: 0
        completed_cases:
          type: integer
          title: Completed Cases
          default: 0
        failed_cases:
          type: integer
          title: Failed Cases
          default: 0
        cancel_requested:
          type: boolean
          title: Cancel Requested
          default: false
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - task_id
      - name
      - benchmark_version_id
      - risk_taxonomy_id
      - status
      title: EvaluationTask
    EvaluationTaskCreateRequest:
      properties:
        name:
          type: string
          title: Name
        benchmark_version_id:
          type: integer
          title: Benchmark Version Id
        risk_taxonomy_id:
          type: integer
          title: Risk Taxonomy Id
        model_bindings:
          items:
            $ref: '#/components/schemas/ModelBinding'
          type: array
          title: Model Bindings
        attack_template_ids:
          items:
            type: integer
          type: array
          title: Attack Template Ids
        judge_profile_ids:
          items:
            type: integer
          type: array
          title: Judge Profile Ids
        rule_set_ids:
          items:
            type: integer
          type: array
          title: Rule Set Ids
        execution:
          $ref: '#/components/schemas/ExecutionConfig'
        scoring:
          $ref: '#/components/schemas/ScoringConfig'
        priority:
          type: string
          title: Priority
          default: normal
      type: object
      required:
      - name
      - benchmark_version_id
      - risk_taxonomy_id
      - model_bindings
      - judge_profile_ids
      - rule_set_ids
      - execution
      - scoring
      title: EvaluationTaskCreateRequest
    ExecutionConfig:
      properties:
        repeat:
          type: integer
          minimum: 1.0
          title: Repeat
          default: 1
        timeout_seconds:
          type: integer
          minimum: 1.0
          title: Timeout Seconds
          default: 120
        retry:
          $ref: '#/components/schemas/RetryConfig'
        concurrency:
          type: integer
          minimum: 1.0
          title: Concurrency
          default: 4
      type: object
      title: ExecutionConfig
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModelBinding:
      properties:
        model_id:
          type: integer
          title: Model Id
        role:
          type: string
          title: Role
          default: target
        model_params:
          additionalProperties: true
          type: object
          title: Model Params
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - model_id
      title: ModelBinding
...
    PaginatedData_AuditLog_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AuditLog'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[AuditLog]
...
    PaginatedData_EvaluationTask_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EvaluationTask'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[EvaluationTask]
...
    PaginatedData_SystemConfig_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SystemConfig'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[SystemConfig]
...
    ScoringConfig:
      properties:
        score_version:
          type: string
          title: Score Version
          default: risk-v1
        weights:
          $ref: '#/components/schemas/ScoringWeights'
      type: object
      title: ScoringConfig
...
    SystemConfig:
      properties:
        config_key:
          type: string
          title: Config Key
        config_value:
          title: Config Value
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_secret:
          type: boolean
          title: Is Secret
          default: false
      type: object
      required:
      - config_key
      - config_value
      title: SystemConfig
    SystemConfigUpdate:
      properties:
        config_value:
          title: Config Value
      type: object
      required:
      - config_value
      title: SystemConfigUpdate
    TaskActionData:
      properties:
        task_id:
          type: integer
          title: Task Id
        status:
          type: string
          title: Status
        cancel_requested:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Cancel Requested
      type: object
      required:
      - task_id
      - status
      title: TaskActionData
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    LocalToken:
      type: apiKey
      in: header
      name: X-Local-Token